extract-linkedin-premium-users-from-linkedin-groups/agentic/agno/pyproject.toml
[project]
name = "agnoAgent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.24",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.39.0"
]
